arulesViz: Visualizing Association Rules and Frequent Itemsets

Extends package 'arules' with various visualization techniques for association rules and itemsets. The package also includes several interactive visualizations for rule exploration. Michael Hahsler (2017) doi:10.32614/RJ-2017-047 .

Main Functions

plot() plot association rules.

inspectDT() for interactive rule tables.

ruleExplorer() for interactive rule inspection with tables and graphs.

See also

Author

Maintainer: Michael Hahsler mhahsler@lyle.smu.edu (ORCID) [copyright holder]

Other contributors:

Giallanza Tyler [contributor]

Sudheer Chelluboina [contributor]