Interactive Scatter Plot for Association Rules using plotly

Plot an interactive scatter plot for association rules using plotly.

plotly_arules(x, method = "scatterplot", measure = c("support", "confidence"),
  shading = "lift", max = 1000, ...)

Arguments

x

an object of class "rules".

method

currently the methods "scatterplot", "two-key plot" and "matrix" are supported.

measure

measure(s) of interestingness (e.g., "support", "confidence", "lift", "order") used in the visualization as x and y-axis.

shading

measure of interestingness used for color shading.

max

client side processing in plotly is expensive. We restrict the number of rules to the max best rules (according to the measure used for shading.)

…

The following additional arguments can be used: colors to specify a color palette, precision to specify the precision used for printing quality measures, and jitter to reduce overplotting in scatterplots (defaults to .1 if overplotting would occur). Further arguments are passed on to plot_ly() as markter attributes (e.g., size, symbol and opacity).

Value

The plotly object (plotly_hash).

Examples

## Not run: ------------------------------------
#  
# library(plotly)
# data(Groceries)
# rules <- apriori(Groceries, parameter=list(support=0.001, confidence=0.8))
# rules
# 
# # interactive scatter plot visualization
# plotly_arules(rules)
# plotly_arules(rules, measure = c("support", "lift"), shading = "confidence")
# plotly_arules(rules, method = "two-key plot")
# 
# # add jitter, change color and markers and add a title
# plotly_arules(rules, jitter = 10, opacity = .7, size = 10, symbol = 1, 
#   colors = c("blue", "green")) <!-- %&gt;% layout(title = &quot;Some Grocery Rules&quot;) -->
# 
# 
# # save a plot as a html page
# p <- plotly_arules(rules)
# htmlwidgets::saveWidget(p, "arules.html", selfcontained = FALSE)
# browseURL("arules.html")
# # Note: selfcontained seems to make the browser slow.
# 
# # interactive matrix visualization
# plotly_arules(rules, method = "matrix") <!-- %&gt;% layout(title = &quot;Reordered Matrix&quot;) -->
## ---------------------------------------------